requirements.txt
python-telegram-bot==21.2
openai==1.54.0
python-dotenv==1.0.1
Pillow==10.4.0
